langgraph_mastery/04_enterprise/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.20
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0
reportlab>=4.0.0
python-json-logger>=2.0.0
